+++ floaty_chatheads_ios/lib/floaty_chatheads_ios.dart
@@ -77,6 +77,8 @@ class FloatyChatheadsIOS extends FloatyChatheadsPlatform {
             .values[config.entranceAnimation.index],
         theme: themeMsg,
         debugMode: config.debugMode,
+        autoLaunchOnBackground: config.autoLaunchOnBackground,
+        persistOnAppClose: config.persistOnAppClose,
       ),
     );
   }